Add the cascade-color style property.

+++ modules/axmol/style/main/include/bim/axmol/style/display_property_flags_fwd.hpp
@@ -4,7 +4,7 @@
 
 namespace bim::axmol::style
 {
-  enum class display_property_flags : std::uint8_t;
+  enum class display_property_flags : std::uint16_t;
 
   display_property_flags operator&(display_property_flags lhs,
                                    display_property_flags rhs);
